Add type declarations for config.ts
