+++ src/example-config.ts
@@ -1,38 +1,4 @@
-import { HexColorString } from "discord.js";
-
-type SelfRoleType = {
-  ID: string;
-  NAME: string;
-  EMOJI: string;
-};
-
-type ColorsType = {
-  SUCCESS: HexColorString;
-  ERROR: HexColorString;
-  WARNING: HexColorString;
-};
-
-export type ConfigType = {
-  BOT_TOKEN: string;
-  MONGODB_URI: string;
-  CLIENT_ID: string;
-  GUILD_ID: string;
-  STAFF_ROLES: string[];
-  ROLE_CHANNEL: string;
-  COLORS: ColorsType;
-  BAN_ROLE: string;
-  BAN_CHANNEL: string;
-  TICKET_CHANNEL: string;
-  TICKET_CATEGORY: string;
-  SPAM_HANDLER: boolean;
-  WARN_EXPIRES: number;
-  WARN_THRESHOLD: number;
-  WARN_BAN_DAYS: number;
-  DELETE_INVITES: boolean;
-  SELF_ROLES: SelfRoleType[];
-};
-
-export const config: ConfigType = {
+export const config = {
   BOT_TOKEN: "BOT_TOKEN",
   MONGODB_URI: "MONGODB_URI",
   CLIENT_ID: "CLIENT_ID",
